saw a missed "self"
not sure those are even necessary, but there it is.
so i was originally thinking we'd have a list of "active" players, and a list of "dead" players, but it seems like that would defeat the purpose of the "status" trait in the player object.  I think it would be better to just have a roster, and the players are alive or dead, and we just skip over dead players when making decisions.

Another option is to have the Roster, which we remove players from when they die, and instead add them to the dead_players list... technically it would be more efficient to remove a player from the list if we're just going to skip over them for being dead, but in reality, unless you have hundreds of thousands of players, a couple extra loop iterations is going to be unnoticeable.

I kind of prefer to have all players always be on the roster, so we have a sort of master list that can remain untouched, and if we find it useful, we can create a dead_player list that is a separate list of only the dead players.