basic save states using pickle
added the ability to save/load the roster using the pickle class.
I think the roster is really all we would need, right? it contains the
complete list of players, and each player object should include
everything we could possibly need to know... we lose our existing dead
player list, but maybe we don't need that for anything anyway (and we
could easily rebuild it upon the loading of our roster)