loads o'changes
Hiddenblade.py:
save file no longer gets timestamped, so only one save file will exist
at a time (seems like we wouldnt need more than 1...)

SMSManager.py:
This could be considered the "main" program to run.  It's basically
serving as a template for now, for future text/email support.  a
function was created to parse a message, and take an action based on
that message.
For now, the actions just result in prints to the console, but in the
future, those prints would be sent to the appropriate recipients via
text.
Eventually, this piece will manage the threads as described within the
file.